refactor(compiler): hoist FlatPlan to its own module
The `tropical_plan_4` schema is a contract between the TS compiler
and the C++ engine; both legacy and resolved-IR emit boundaries
produce it. Living in `flatten.ts` means every consumer importing
the type pulls in `flatten.ts` even when it doesn't use any of its
functions — a footgun once `flatten.ts` is retired in the upcoming
cleanup PR.

Move `interface FlatPlan` to `compiler/flat_plan.ts`. `flatten.ts`
re-exports the type for the few in-tree callers we don't migrate
this commit. Eight type-only import sites updated to point at the
new home directly:

- compiler/emit_wasm.ts, emit_wasm.test.ts, wasm_runtime.test.ts,
  wasm_vs_jit_equiv.test.ts, web_plans_vs_jit.test.ts
- compiler/ir/compile_resolved.ts, compile_session.ts
- scripts/eval_program.ts
- web/host/compiler.ts

No behavior change. Tests: 929 pass, 1 skip, 1 local-only fail (the
same untracked schema_audit noise).
