+++ compiler/emit_wasm.test.ts
@@ -6,7 +6,7 @@
  */
 
 import { describe, test, expect } from 'bun:test'
-import type { FlatPlan } from './flatten'
+import type { FlatPlan } from './flat_plan'
 import type { NInstr, NOperand, ScalarType } from './emit_numeric'
 import { emitWasm } from './emit_wasm'
